backend/requirements.txt
fastapi==0.121.1
uvicorn==0.38.0
python-dotenv==1.2.1

# ML stack
torch==2.2.2
scikit-learn==1.7.2
scipy==1.15.3
numpy==1.26.4
sentence-transformers==5.1.2
transformers==4.57.1
tokenizers==0.22.1
safetensors==0.6.2
joblib==1.5.2

# Text & utils
nltk==3.9.1
regex==2025.11.3
tqdm==4.67.1

# Pinecone & vector search
pinecone==7.3.0

# Optional: only if you use PDFs
PyPDF2==3.0.1

# Optional: general utils
pandas==2.3.3
requests==2.32.5
